created an interface for the Iterator
