+++ core/test/com/group66/game/cannon/BallTest.java
@@ -8,6 +8,7 @@ import org.junit.Test;
 
 import com.badlogic.gdx.math.Circle;
 import com.group66.game.cannon.BallType;
+import com.group66.game.cannon.ballgraph.BallGraph;
 
 /**
  * The Class BallTest.
